src/app/dashboard/dashboard.component.ts
import { Component, OnInit } from '@angular/core';
import { Reservation } from '../reservation';
import { ReservationService } from '../reservation.service';

@Component({
  selector: 'app-dashboard',
  templateUrl: './dashboard.component.html',
  styleUrls: [ './dashboard.component.css' ]
})
export class DashboardComponent implements OnInit {
  heroes: Reservation[] = [];

  constructor(private heroService: ReservationService) { }

  ngOnInit() {
    this.getHeroes();
  }

  getHeroes(): void {
    this.heroService.getReservations()
      .subscribe(heroes => this.heroes = heroes.slice(1, 5));
  }
}
